Use `find_package()` to locate dependencies
This changes CMakeLists.txt in the src and test directories so that
instead of looking for the Conan-provided dependencies, CMake searches
the host system for the dependencies. Since the conan_cmake wrapper was
set up as described in the docs, it can provide these dependencies
properly. As a side effect, these CMakeLists.txt files are no longer
dependent on Conan.